backend/requirements.txt
# Django and Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==2.3.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1

# File Handling
Pillow==10.1.0
django-storages==1.14.2

# Validation & Serialization
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Development
django-debug-toolbar==4.2.0

# Environment & Configuration
python-decouple==3.8 django-filter>=23.0
